Fix paths and structure on CDS and US Census
