+++ task_geo/data_sources/__init__.py
@@ -1,7 +1,11 @@
+from task_geo.data_sources.covid.cds import cds
+from task_geo.data_sources.demographics.us_census import us_census
 from task_geo.data_sources.noaa import noaa_api
 
 AVAILABLE_DATA_SOURCES = {
     'noaa_api': noaa_api,
+    'cds': cds,
+    'us_census': us_census
 }
